Backed out 6 changesets (bug 1622530, bug 1622561) for spidermonkey failure at /builds/worker/workspace/build/src/js/src/frontend/SourceNotes.h:282:31
Backed out changeset 4e13790492c6 (bug 1622530)
Backed out changeset 4ce323a41651 (bug 1622530)
Backed out changeset e72b6cf947cd (bug 1622530)
Backed out changeset fc2ad74a76a6 (bug 1622530)
Backed out changeset 4249e5f9f387 (bug 1622530)
Backed out changeset fabc22ae4259 (bug 1622561)

UltraBlame original commit: 7439fe7f29557caecc518bc6726cf8c8c3c29b7d

+++ js/src/frontend/BytecodeSection.cpp
@@ -140,8 +140,7 @@ bool CGTryNoteList::append(TryNoteKind kind, uint32_t stackDepth,
   
   
 
-  TryNote note(uint32_t(kind), stackDepth, start.toUint32(),
-               (end - start).toUint32());
+  TryNote note(kind, stackDepth, start.toUint32(), (end - start).toUint32());
 
   return list.append(note);
 }
